fix dtype bug
It would raise error if we use the bfloat16 dtype.

+++ src/transformers/pipelines/text_classification.py
@@ -202,7 +202,7 @@ class TextClassificationPipeline(Pipeline):
                 function_to_apply = ClassificationFunction.NONE
 
         outputs = model_outputs["logits"][0]
-        outputs = outputs.numpy()
+        outputs = outputs.float().numpy()
 
         if function_to_apply == ClassificationFunction.SIGMOID:
             scores = sigmoid(outputs)
